requirements.txt
numpy==1.14.2
opencv-python==3.4.0.12
Pillow==5.1.0
setuptools==38.2.5
